refactor(handlers): extract activate.ts handler from intent.ts (Stage B)
handleSAPActivate + batch-activation status helpers. The batch helpers are
shared with handleSAPWrite (batch_create activateAtEnd), so activate.ts
exports them and intent.ts imports them back (acyclic). 3672 tests, A1 zero-diff.
